refactor: added index.ts for components, replaced <> with Fragment
