+++ src/components/AuthPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AuthPage';
