+++ src/components/Logo/Logo.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import logo from '../assets/images/logo.png';
-import logoLight from '../assets/images/logo-light.png';
+import logo from '../../assets/images/logo.png';
+import logoLight from '../../assets/images/logo-light.png';
 
 interface LogoProps {
   variant?: string;
