+++ src/components/MapPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MapPage';
