+++ src/components/ProfilePage/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ProfilePage';
