+++ src/index.scss
@@ -1,5 +1,3 @@
-// @import-normalize;
-
 *,
 *::before,
 *::after {
